[LOR-209] - refactoring packages, file and structs name
